Move functions to DOMUpdates object
